+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/file/MiniAppFilePicker.kt
@@ -1,20 +1,50 @@
 package com.rakuten.tech.mobile.miniapp.file
 
+import android.app.Activity
+import android.content.Context
 import android.net.Uri
 import android.webkit.ValueCallback
+import android.webkit.WebChromeClient
+import com.rakuten.tech.mobile.miniapp.MiniAppFilePickingException
 
-/** The file picker of a miniapp. **/
-interface MiniAppFilePicker {
+/**
+ * The file picker of a miniapp which needs to be implemented by the HostApp.
+ * @param requestCode of file picking using an intent inside sdk, which will also be used
+ * to retrieve the data by [Activity.onActivityResult] in the HostApp.
+ **/
+open class MiniAppFilePicker(var requestCode: Int) {
+
+    internal var callback: ValueCallback<Array<Uri>>? = null
 
     /**
-     * Request to pick a file from the device within the miniapp webview per [requestCode].
-     * @param filePathCallback a file path callback to pass to HostApp has been found
-     * by [MiniAppWebChromeClient.onShowFileChooser].
-     * @param callback to pass the request code of file picking using an intent inside sdk,
-     * which will also be used to retrieve the data using onActivityResult in the host app.
+     * HostApp can override [WebChromeClient.onShowFileChooser] to choose a file
+     * which is requested using HTML forms with 'file' input type.
+     * @param filePathCallback The callback to receive the Uris.
+     * @param fileChooserParams The file chooser parameters can be used to create an intent to choose file.
+     * @param context Activity's context from the HostApp.
      */
-    fun requestFile(
+    @Suppress("TooGenericExceptionCaught", "SwallowedException")
+    fun onShowFileChooser(
         filePathCallback: ValueCallback<Array<Uri>>?,
-        callback: (requestCode: Int) -> Unit
-    )
+        fileChooserParams: WebChromeClient.FileChooserParams?,
+        context: Context
+    ): Boolean {
+        try {
+            this.callback = filePathCallback
+            val intent = fileChooserParams?.createIntent()
+            (context as Activity).startActivityForResult(intent, requestCode)
+        } catch (e: Exception) {
+            throw (MiniAppFilePickingException(e.message.toString()))
+        }
+        return true
+    }
+
+    /**
+     * Receive the files from the HostApp.
+     * @param files The array of Uri to be invoked by a filePathCallback after successfully retrieved
+     * by [Activity.onActivityResult] in the HostApp.
+     */
+    fun onReceivedFiles(files: Array<Uri>) {
+        callback?.onReceiveValue(files)
+    }
 }